pulled randomizers out of gender_generator
